WIP: Players position sync.

+++ src/QuestForTheCrownJ/qftcj-server/ServerStructs.h
@@ -3,8 +3,8 @@
 
 typedef struct s_logged_user
 {
-	int user_id;
-	int player_id;
+	long long user_id;
+	long long player_id;
 	std::string map_name;
 	std::shared_ptr<qfcbase::Entity> game_entity;
 	// last command date time
